requirements.txt
yfinance>=0.2.36
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0
scipy>=1.12.0
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=2.3.0
alpha_vantage>=3.0.0
